add docker changes: initialize based on cloudhost node - 4th parameter pass cloudhost node to executer service allow non-cloudhost urls to execute

+++ cloudhost-src/src/main/java/org/cdlib/mrt/cloudhost/app/jersey/JerseyUtil.java
@@ -87,10 +87,10 @@ public class JerseyUtil
      * @return node identifier
      * @throws TException invalid node identifier
      */
-    protected int getNodeID(String parm)
+    protected long getNodeID(String parm)
         throws TException
     {
-        int retNode = 0;
+        long retNode = 0;
         if (StringUtil.isAllBlank(parm)){
             throw new TException.INVALID_OR_MISSING_PARM("Node not supplied");
         }
